requirements.txt
scipion-pyworkflow
scipy
numpy
aicsimageio